requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
psycopg[binary]>=3.1.0
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.22
python-dotenv==1.2.1
email-validator==2.3.0
PyJWT==2.8.0
jinja2==3.1.6
filetype==1.2.0
pikepdf>=9.0.0
Pillow>=12.0.0
img2pdf>=0.6.0
httpx>=0.27.0
bcrypt>=4.0.0
apscheduler==3.10.4
